refactor: factor cube-cavity NumPy mesh primitives into reference/numpy/mesh.py (#103)
The two NumPy cube-cavity entry points (cube_cavity.py from #92 and
cube_cavity_minimal.py from #93) shipped with duplicated implementations
of cube_tet_mesh and cube_interior_mask. Factor these out (along with
load_msh / write_msh) into a shared reference/numpy/mesh.py module so
both entry points share one source of truth.

Approach
========

Option A from the issue (factor) over Option B (merge). Reasons:

- The two entry points have *different* end-to-end APIs: cube_cavity.py
  exports run_cube_cavity returning a dict keyed by `n_int` /
  `k_int_frobenius` etc. for the n=10 Gmsh-fixture path, while
  cube_cavity_minimal.py exports solve_cube_cavity returning a dict
  keyed by `n_dofs_interior` / `k_diag_sum` for the programmatic n=4
  path the JAX/TF-Java siblings consume. Merging would force one path
  to break or carry both APIs side-by-side; factoring the shared mesh
  primitives gives clean separation without disturbing either API.
- reference/jax/cube_cavity.py imports cube_tet_mesh and
  cube_interior_mask from cube_cavity_minimal; the re-export keeps the
  JAX import path unchanged.
- reference/jax/gen_cube_cavity_fixture.py imports solve_cube_cavity
  from cube_cavity_minimal; that lives in cube_cavity_minimal still
  and is not affected.
- reference/numpy/gen_cube_cavity_baseline.py imports cube_tet_mesh /
  write_msh from cube_cavity; re-exports keep that path working too.

Verification
============

- Smoke-tested in-process: cube_tet_mesh is the *same* function object
  when imported from cube_cavity, cube_cavity_minimal, or mesh — proving
  the shared-source factoring landed correctly.
- Replicated the exact JAX import lines: `from cube_cavity_minimal
  import cube_interior_mask, cube_tet_mesh` and `from
  cube_cavity_minimal import solve_cube_cavity` — both still resolve.
- `cargo test -p geode-validation --release --test
  cube_cavity_numpy_reference` and `cargo test -p geode-validation
  --release --test cube_cavity_jax_reference`: same pass/fail
  composition as on origin/main. The
  `cube_cavity_burn_matches_numpy_reference_at_all_substages` test was
  already failing on origin/main (Burn `m_int_diag` drift ~2.5e-10 vs
  tolerance 1e-14, a backend-precision artifact unrelated to this
  Python refactor) — verified by checking out origin/main's reference/
  tree under the same fixtures and reproducing the same failure.

Closes #103
